+++ org-vscode/test/unit/run.js
@@ -36,6 +36,7 @@ const tests = [
   require(path.join(__dirname, 'column-view.test.js')),
   require(path.join(__dirname, 'html-utils-browser.test.js')),
   require(path.join(__dirname, 'org-preview-export.test.js')),
+  require(path.join(__dirname, 'org-link-targets.test.js')),
   require(path.join(__dirname, 'year-executive-report.test.js')),
   require(path.join(__dirname, 'year-summary-tags.test.js')),
   require(path.join(__dirname, 'link-type-registry.test.js')),